Update of the shape file - setup for age, gender, safety and security
This version of the model includes corrections in the number of people by age brakets and gender. Whe had the wrong column of percentages applied to the number of males and females. Max and I made chenges in the shape file.
Also, this one includes the assigment of age and gender by communes in the initialization tab.
 Finally, includes the safety and security tabs to have accidents and insecuiryt incidents at each tick. This procedure depends on the transport type which is not included yet in this version of the model. We need to combine it with the Richard's version.